+++ samanthas_telegram_bot/api_clients/__init__.py
@@ -1,0 +1,5 @@
+from samanthas_telegram_bot.api_clients.backend.backend_client import BackendClient
+
+__all__ = [
+    "BackendClient",
+]
