+++ samanthas_telegram_bot/api_clients/backend/backend_client.py
@@ -6,7 +6,7 @@ import httpx
 from telegram import Update
 from telegram.constants import ParseMode
 
-from samanthas_telegram_bot.api_queries.auxil.constants import (
+from samanthas_telegram_bot.api_clients.auxil.constants import (
     API_URL_CHECK_EXISTENCE_OF_CHAT_ID,
     API_URL_CHECK_EXISTENCE_OF_PERSONAL_INFO,
     API_URL_ENROLLMENT_TEST_GET_LEVEL,
@@ -20,13 +20,11 @@ from samanthas_telegram_bot.api_queries.auxil.constants import (
     API_URL_YOUNG_TEACHERS_LIST_CREATE,
     DataDict,
 )
-from samanthas_telegram_bot.api_queries.auxil.enums import LoggingLevel
-from samanthas_telegram_bot.api_queries.auxil.exceptions import (
-    BackendClientError,
-    BaseApiClientError,
-)
-from samanthas_telegram_bot.api_queries.auxil.models import NotificationParams
-from samanthas_telegram_bot.api_queries.base_api_client import BaseApiClient
+from samanthas_telegram_bot.api_clients.auxil.enums import LoggingLevel
+from samanthas_telegram_bot.api_clients.auxil.models import NotificationParams
+from samanthas_telegram_bot.api_clients.backend.exceptions import BackendClientError
+from samanthas_telegram_bot.api_clients.base.base_api_client import BaseApiClient
+from samanthas_telegram_bot.api_clients.base.exceptions import BaseApiClientError
 from samanthas_telegram_bot.auxil.escape_for_markdown import escape_for_markdown
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
