+++ samanthas_telegram_bot/api_clients/backend/exceptions.py
@@ -1,0 +1,5 @@
+from samanthas_telegram_bot.api_clients.base.exceptions import BaseApiClientError
+
+
+class BackendClientError(BaseApiClientError):
+    pass
