+++ samanthas_telegram_bot/api_clients/base/base_api_client.py
@@ -5,14 +5,14 @@ import httpx
 from httpx import Response
 from telegram import Update
 
-from samanthas_telegram_bot.api_queries.auxil.constants import (
+from samanthas_telegram_bot.api_clients.auxil.constants import (
     MAX_ATTEMPTS_TO_GET_DATA_FROM_API,
     TIMEOUT_IN_SECS_BETWEEN_API_REQUEST_ATTEMPTS,
     DataDict,
 )
-from samanthas_telegram_bot.api_queries.auxil.enums import HttpMethod, LoggingLevel
-from samanthas_telegram_bot.api_queries.auxil.exceptions import BaseApiClientError
-from samanthas_telegram_bot.api_queries.auxil.models import NotificationParamsForStatusCode
+from samanthas_telegram_bot.api_clients.auxil.enums import HttpMethod, LoggingLevel
+from samanthas_telegram_bot.api_clients.auxil.models import NotificationParamsForStatusCode
+from samanthas_telegram_bot.api_clients.base.exceptions import BaseApiClientError
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.data_structures.constants import CALLER_LOGGING_STACK_LEVEL
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
