+++ samanthas_telegram_bot/api_clients/base/exceptions.py
@@ -1,0 +1,2 @@
+class BaseApiClientError(Exception):
+    pass
