+++ samanthas_telegram_bot/api_clients/smalltalk.py
@@ -9,14 +9,14 @@ from dotenv import load_dotenv
 from telegram import Update
 from telegram.constants import ParseMode
 
-from samanthas_telegram_bot.api_queries.auxil.constants import (
+from samanthas_telegram_bot.api_clients.auxil.constants import (
     MAX_ATTEMPTS_TO_GET_DATA_FROM_API,
     SMALLTALK_URL_GET_RESULTS,
     SMALLTALK_URL_GET_TEST,
     TIMEOUT_IN_SECS_BETWEEN_API_REQUEST_ATTEMPTS,
 )
-from samanthas_telegram_bot.api_queries.auxil.enums import LoggingLevel
-from samanthas_telegram_bot.api_queries.auxil.exceptions import BaseApiClientError
+from samanthas_telegram_bot.api_clients.auxil.enums import LoggingLevel
+from samanthas_telegram_bot.api_clients.base.exceptions import BaseApiClientError
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.data_structures.constants import ALL_LEVELS
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
