--- samanthas_telegram_bot/api_queries/__init__.py
@@ -1,1 +1,0 @@
-
