--- samanthas_telegram_bot/api_queries/auxil/exceptions.py
@@ -1,6 +1,0 @@
-class BaseApiClientError(Exception):
-    pass
-
-
-class BackendClientError(BaseApiClientError):
-    pass
