+++ tests/api_requests/test_smalltalk.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-import samanthas_telegram_bot.api_queries.smalltalk as smalltalk
+import samanthas_telegram_bot.api_clients.smalltalk as smalltalk
 from samanthas_telegram_bot.data_structures.enums import SmalltalkTestStatus
 from tests.constants import TEST_DIR
